+++ common/src/main/java/net/merchantpug/apugli/mixin/xplatform/common/ItemStackMixin.java
@@ -1,15 +1,12 @@
 package net.merchantpug.apugli.mixin.xplatform.common;
 
 import net.merchantpug.apugli.platform.Services;
-import net.merchantpug.apugli.power.EdibleItemPower;
 import net.merchantpug.apugli.registry.power.ApugliPowers;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
-import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.UseAnim;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
@@ -19,8 +16,6 @@ import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
-
-import java.util.Optional;
 
 @Mixin(ItemStack.class)
 public abstract class ItemStackMixin {
@@ -69,38 +64,6 @@ public abstract class ItemStackMixin {
                 power.executeIncreaseAction((ItemStack)(Object)this);
             }
         });
-    }
-
-    @Inject(method = "getUseAnimation", at = @At("HEAD"), cancellable = true)
-    private void apugli$getUseAction(CallbackInfoReturnable<UseAnim> cir) {
-        ItemStack stack = (ItemStack)(Object)this;
-        if (!(Services.PLATFORM.getEntityFromItemStack(stack) instanceof LivingEntity living)) return;
-        Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(living.level(), stack)).findFirst();
-        power.ifPresent(edibleItemPower -> cir.setReturnValue(edibleItemPower.getUseAction().equals(EdibleItemPower.EatAnimation.DRINK) ? UseAnim.DRINK : UseAnim.EAT));
-    }
-
-    @Inject(method = "getUseDuration", at = @At("HEAD"), cancellable = true)
-    private void apugli$getMaxUseTime(CallbackInfoReturnable<Integer> cir) {
-        ItemStack stack = (ItemStack)(Object)this;
-        if (!(Services.PLATFORM.getEntityFromItemStack(stack) instanceof LivingEntity living)) return;
-        Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(living.level(), stack)).findFirst();
-        power.ifPresent(edibleItemPower -> cir.setReturnValue(edibleItemPower.getFoodComponent().isFastFood() ? 16 : 32));
-    }
-
-    @Inject(method = "getDrinkingSound", at = @At("HEAD"), cancellable = true)
-    private void apugli$getDrinkSound(CallbackInfoReturnable<SoundEvent> cir) {
-        ItemStack stack = (ItemStack)(Object)this;
-        if (!(Services.PLATFORM.getEntityFromItemStack(stack) instanceof LivingEntity living)) return;
-        Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(living.level(), stack) && p.getSound() != null).findFirst();
-        power.ifPresent(edibleItemPower -> cir.setReturnValue(edibleItemPower.getSound()));
-    }
-
-    @Inject(method = "getEatingSound", at = @At("HEAD"), cancellable = true)
-    private void apugli$getEatSound(CallbackInfoReturnable<SoundEvent> cir) {
-        ItemStack stack = (ItemStack)(Object)this;
-        if (!(Services.PLATFORM.getEntityFromItemStack(stack) instanceof LivingEntity living)) return;
-        Optional<EdibleItemPower> power = Services.POWER.getPowers(living, ApugliPowers.EDIBLE_ITEM.get()).stream().filter(p -> p.doesApply(living.level(), stack) && p.getSound() != null).findFirst();
-        power.ifPresent(edibleItemPower -> cir.setReturnValue(edibleItemPower.getSound()));
     }
 
 }