+++ common/src/main/java/net/merchantpug/apugli/registry/condition/ApugliItemConditions.java
@@ -4,7 +4,9 @@ import net.merchantpug.apugli.condition.factory.IConditionFactory;
 import net.merchantpug.apugli.condition.factory.item.BaseEnchantmentCondition;
 import net.merchantpug.apugli.condition.factory.item.OnCooldownCondition;
 import net.merchantpug.apugli.platform.Services;
+import net.minecraft.util.Tuple;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 
 public class ApugliItemConditions {
     
@@ -13,7 +15,7 @@ public class ApugliItemConditions {
         register("on_cooldown", new OnCooldownCondition());
     }
     
-    private static void register(String name, IConditionFactory<ItemStack> action) {
+    private static void register(String name, IConditionFactory<Tuple<Level, ItemStack>> action) {
         Services.CONDITION.registerItem(name, action);
     }