+++ fabric/src/main/java/net/merchantpug/apugli/access/FactoryInstanceAccess.java
@@ -1,8 +1,0 @@
-package net.merchantpug.apugli.access;
-
-import com.google.gson.JsonElement;
-
-public interface FactoryInstanceAccess {
-    JsonElement apugli$getJson();
-    void apugli$setJson(JsonElement json);
-}