+++ fabric/src/main/java/net/merchantpug/apugli/mixin/fabric/common/ActionTypeMixin.java
@@ -4,7 +4,7 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import io.github.apace100.apoli.power.factory.action.ActionFactory;
 import io.github.apace100.apoli.power.factory.action.ActionType;
-import io.github.apace100.apoli.util.NamespaceAlias;
+import io.github.apace100.apoli.util.IdentifierAlias;
 import net.merchantpug.apugli.Apugli;
 import net.merchantpug.apugli.action.factory.entity.CustomProjectileAction;
 import net.merchantpug.apugli.util.TextureUtil;
@@ -18,14 +18,16 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
+import java.util.Optional;
+
 @Mixin(value = ActionType.class, remap = false)
 public class ActionTypeMixin<T> {
 
     @Shadow @Final private String actionTypeName;
 
     @Inject(method = "read(Lcom/google/gson/JsonElement;)Lio/github/apace100/apoli/power/factory/action/ActionFactory$Instance;", at = @At(value = "RETURN"), locals = LocalCapture.CAPTURE_FAILHARD, remap = false)
-    private void apugli$handleActions(JsonElement jsonElement, CallbackInfoReturnable<ActionFactory<T>.Instance> cir, JsonObject jsonObject, String typeIdentifier, ResourceLocation type) {
-        if (this.actionTypeName.equals("EntityAction") && (type.equals(Apugli.asResource("custom_projectile")) || NamespaceAlias.hasAlias(type) && NamespaceAlias.resolveAlias(type).equals(Apugli.asResource("custom_projectile")))) {
+    private void apugli$handleActions(JsonElement jsonElement, CallbackInfoReturnable<ActionFactory<T>.Instance> cir, JsonObject jsonObject, ResourceLocation type, Optional actionFactory) {
+        if (this.actionTypeName.equals("EntityAction") && (type.equals(Apugli.asResource("custom_projectile")) || IdentifierAlias.hasAlias(type) && IdentifierAlias.resolveAlias(type).equals(Apugli.asResource("custom_projectile")))) {
             if (jsonObject.has("texture_url")) {
                 String url = GsonHelper.getAsString(jsonObject, "texture_url");
                 ResourceLocation textureLocation = null;