+++ fabric/src/main/java/net/merchantpug/apugli/network/ApugliPackets.java
@@ -57,7 +57,6 @@ public class ApugliPackets {
             ClientPlayNetworking.registerReceiver(ExecuteBiEntityActionClientPacket.ID, createS2CHandler(ExecuteBiEntityActionClientPacket::decode, ExecuteBiEntityActionClientPacket::handle));
             ClientPlayNetworking.registerReceiver(SyncScalePacket.ID, createS2CHandler(SyncScalePacket::decode, SyncScalePacket::handle));
             ClientPlayNetworking.registerReceiver(MarkLerpedScaleReadyPacket.ID, createS2CHandler(MarkLerpedScaleReadyPacket::decode, MarkLerpedScaleReadyPacket::handle));
-            ClientPlayNetworking.registerReceiver(ModifyEnchantmentLevelPacket.ID, createS2CHandler(ModifyEnchantmentLevelPacket::decode, ModifyEnchantmentLevelPacket::handle));
         });
     }