+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 - Ported the mod to 1.20.2.
+
+- Aliased `apugli:edible_item` to `apoli:edible_item`.
+- Aliased `apugli:modify_enchantment_level` to `apoli:modify_enchantment_level`.